chore: remove unused cors configuration
CORS configurations are specific to browser -> server communications.
App gateway is not addressable on the internet and so should never be
callable by a browser.

Likely this was originally added to test using an in-browser tool like
postman. Instead, use the postman native app or curl, such as:

```
curl -H "Content-Type: application/json" \
     --data @body.json \
     http://localhost:5000/a/b/c
```

... where @body.json is the name of a file 'body.json' containing the
post payload.