ci: expose version, revision, and created docker labels to the image itself as env vars
When a docker image is created, we label it with a number of
standardized image labels. In order to allow the application to report
its identity, we will pass three of these labels in as environment
variables.

The three labels, as defined by the spec are:

 - org.opencontainers.image.created date and time on which the image was
   built (string, date-time as defined by RFC 3339).
 - org.opencontainers.image.version version of the packaged software
 - org.opencontainers.image.revision Source control revision identifier
   for the packaged software.

We source this information from the output of our github
docker/metadata-action@v5. The result of which sets 'revision' to the
git
sha of the commit built and 'version' to the assigned git tag, a
pull-request label (eg. pr-4), a branch name, or 'unknown'.

See
https://specs.opencontainers.org/image-spec/annotations/#pre-defined-annotation-keys